Add TestExportCertificateAsPemStr test case
